+++ CHANGELOG.md
@@ -1,3 +1,25 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.0.1] - 2024-09-05
+
+### Added
+
+- Macro wip
+- Macro DefaultInit
+- Macro debug_print
+- Macro CustomCodableKey
+- Macro CustomDefault
+- Macro CustomDate
+- Macro CustomCodable
+- Macro CustomHashable
+- Macro CustomEquatable
+- Macro StringOrInt
+- Macro StringToDouble
+- Macro CustomURL
+- Macro AutoMockable
+- Macro NodePrinter